+++ bundles/framework/maplegend/publisher/MapLegendToolComponent.jsx
